rename compress.rs to decompress.rs
