+++ src/lib.rs
@@ -9,6 +9,7 @@ pub mod constant_strings_paths;
 pub mod content_window;
 pub mod copy_move;
 pub mod cryptsetup;
+pub mod decompress;
 pub mod event_dispatch;
 pub mod event_exec;
 pub mod fileinfo;
